requirements.txt
fastapi ==  0.75.1
uvicorn == 0.17.6
python-multipart == 0.0.5
tortoise-orm == 0.19.0
passlib == 1.7.4
bcrypt == 3.2.2
PyJWT == 2.3.0